Core: Have the recompiler just deal with the Program Counter as 32bit

+++ Source/Project64-core/N64System/Recompiler/ExitInfo.h
@@ -33,7 +33,7 @@ struct CExitInfo
 
     std::string Name;
     uint32_t ID;
-    uint64_t TargetPC;
+    uint32_t TargetPC;
     CRegInfo ExitRegSet;
     ExitReason Reason;
     PIPELINE_STAGE PipelineStage;
